AP: turn on announcement attempt solicited
Note that we are almost capable of providing virtual ports, though
currently we do not utilize that aspect of Linux kernel MACsec
implementations. So even in the case we receive the EAPOL-START or
EAPOL-Announcement-Req with NID set tlvs, we just keep transmitting
generic announcement. See IEEE 802.1X-2010 10.2 paragraph just above
NOTE 3, p.p. 84.